Add image job handling, Azure Blob integration,
Introduced a system for managing image generation jobs:
- Added migrations for database schema changes,
- Implemented `JobPollingService` to process jobs, upload images to Azure Blob Storage, and update the database.
- Added `AzureBlobStorageService` with methods for uploading files and generating SAS URLs.
- Defined interfaces (`IImageService`, `IStorageService`) for API and storage interactions.
- Improved code organization with new namespaces and updated the database context snapshot.

+++ firefly/Models/JobResult.cs
@@ -38,7 +38,7 @@ namespace firefly.Models
     public class FireflyImageOutput
     {
         [JsonPropertyName("seed")]
-        public long Seed { get; set; }
+        public int Seed { get; set; }
 
         [JsonPropertyName("image")]
         public FireflyImageUrl Image { get; set; }
